Large QueryParams and Request* classes refactor.
Before, both QueryParams and WebserviceParams could be instantiated. The latter wrapped (and delegated to) the former but added lots of methods as well, leading to a very large, messy class API that was not at all specific to the current user request.

WebserviceParams now just a utility class that works on values from QueryParams. QueryParams still contains all parameters from the request, but these are quickly converted to specific Request* objects that only contain values relevant to the specific request. These Request* classes can still be records with many, many fields, but that matches the BlackLab Server API, for better or worse.

+++ engine/src/main/java/nl/inl/blacklab/search/results/SampleParameters.java
@@ -18,11 +18,11 @@ public class SampleParameters {
         return new SampleParameters(true, percentageOfHits, 0, false, 0);
     }
     
-    public static SampleParameters fixedNumber(int numberOfHitsSet, long seedValueSet) {
+    public static SampleParameters fixedNumber(long numberOfHitsSet, long seedValueSet) {
         return new SampleParameters(false, 0, numberOfHitsSet, true, seedValueSet);
     }
     
-    public static SampleParameters fixedNumber(int numberOfHitsSet) {
+    public static SampleParameters fixedNumber(long numberOfHitsSet) {
         return new SampleParameters(false, 0, numberOfHitsSet, false, 0);
     }
